ui-server/requirements.txt
pymongo
apscheduler
requests
flask
flask-classy
hazelcast-python-client
flask-cors
kafka-python
docker